move decompressor to internal
